requirements/base.txt
dj-database-url==1.3.0
Django==4.2
djangorestframework==3.14.0
python-dotenv==1.0.0
psycopg2==2.9.6
django-mptt==0.14.0
django-mptt-admin==2.5.0
django-soft-delete==0.9.21
Pillow==9.5.0